Project HMS

+++ src/main/java/com/HMS/repository/PatientAppointmentRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.HMS.entity.Patient;
 import com.HMS.entity.PatientAppointment;
 
-public interface PatientAppointmentRepo extends JpaRepository<PatientAppointment,Long>
+public interface PatientAppointmentRepo extends JpaRepository<PatientAppointment,Integer>
 {
 
 }
